ml/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
async-timeout==4.0.3
# Commenting out bcc package as version 0.18.0 isn't available on PyPI
# bcc==0.18.0
blinker==1.6.2
# Commenting out system-specific packages that might cause installation issues
# Brlapi==0.8.3
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==3.4.1
# chrome-gnome-shell==0.0.0
click==8.1.3
colorama==0.4.4
# command-not-found==0.3
cryptography==3.4.8
# cupshelpers==1.0
# dbus-python==1.2.18
# defer==1.0.6
distro==1.7.0
exceptiongroup==1.2.2
flask==3.0.2
flask-cors==4.0.0
greenlet==3.1.1
# guake==3.10
h11==0.14.0
# hidpidaemon==18.4.6
httpcore==1.0.7
httplib2==0.20.2
httpx==0.28.1
idna==3.3
importlib-metadata==4.6.4
# jeepney==0.7.1
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
# kernelstub==3.1.4
# keyring==23.5.0
langchain==0.3.21
langchain-core==0.3.48
langchain-text-splitters==0.3.7
langsmith==0.3.18
# language-selector==0.1
# launchpadlib==1.10.16
# lazr.restfulclient==0.14.4
# lazr.uri==1.0.6
# louis==3.20.0
# macaroonbakery==1.3.1
more-itertools==8.10.0
nba_api==1.8.0
# netaddr==0.8.0
# netifaces==0.11.0
numpy==1.26.4
oauthlib==3.2.0
ollama==0.4.7
openai==1.68.2
orjson==3.10.16
packaging==24.2
pandas==2.2.3
# pop-transition==1.1.2
# protobuf==3.12.4
# pycairo==1.20.1
# pycups==2.0.1
pydantic==2.10.6
pydantic_core==2.27.2
# pydbus==0.6.0
# PyGObject==3.42.1
PyJWT==2.3.0
# pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
# python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
# python-debian==0.1.43+ubuntu1.1
# python-gnupg==0.4.8
# python-xlib==0.29
pytz==2022.1
# pyxdg==0.27
PyYAML==6.0.1
# repolib==2.2.1
# repoman==1.4.0
requests==2.32.3
requests-toolbelt==1.0.0
# screen-resolution-extra==0.0.0
# SecretStorage==3.3.1
# sessioninstaller==0.0.0
setuptools-scm==6.4.2
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.39
# systemd-python==234
tenacity==9.0.0
tomli==1.2.2
tqdm==4.67.1
typing_extensions==4.13.0
tzdata==2025.2
# ubuntu-drivers-common==0.0.0
# ufw==0.36.1
urllib3==1.26.5
# wadllib==1.3.6
# xdg==5
# xkit==0.0.0
zipp==1.0.0
zstandard==0.23.0
gunicorn==21.2.0
